DONE convert all the test cases to jest
