+++ server/.babelrc
@@ -1,0 +1,17 @@
+{
+    "presets": [
+      ["env", {
+        "modules": false,
+        "targets": {
+          "browsers": ["> 1%", "last 2 versions", "not ie <= 8"]
+        }
+      }]
+    ],
+    "env": {
+      "test": {
+        "presets": [
+          ["env", { "targets": { "node": "current" }}]
+        ]
+      }
+    }
+  }
