+++ server/app.js
@@ -52,7 +52,10 @@ process.on('unhandledRejection', unhandledRejectionHandler);
 
 //server setup
 const port = process.env.PORT || 80;
-app.listen(port);
-console.log('Server listening on:', port);
+
+if (process.env.NODE_ENV !== 'test') {
+    app.listen(port);
+}
+// console.log('Server listening on:', port);
 
 export default app;
