+++ server/src/Middleware/unhandledRejectionHandler.js
@@ -1,3 +1,2 @@
 export const unhandledRejectionHandler = (err) => {
-  console.log(err);
 };
