+++ server/src/Services/stripe/stripeCustomer.js
@@ -27,8 +27,6 @@ export const CreateCustomer = async (req, res) => {
   //save stripe id to our own db
   let result = await createCustomerModel(customer, email);
 
-  console.log(result);
-
   //send jwt token for user auth requests
   let token = setToken(userId);
