+++ server/src/__tests__/todos.test.js
@@ -1,23 +1,20 @@
 //Require the dev-dependencies
-import chai from 'chai';
-const expect = chai.expect;
-import server from '../../app.js';
 import supertest from 'supertest';
+import server from '../../app.js';
+import expect from 'expect';
 const request = supertest(server)
 
 describe('GET Todo info /get/todos', () => {
     it('get todo info with app id', async () => {
-        request
+        let res = await request
             .get('/api/get/todos?app_id=6025e9b0bcfcf3098888eadc')
-            .end((err, res) => {
-                expect(res.status).to.equal(200)
-            })
+        expect(res.status).toEqual(200)
     });
 })
 
 describe('POST API Todo /post/todo', () => {
     it('create new todo', async () => {
-        request
+        let res = await request
             .post('/api/post/todo')
             .send({
                 title: 'guribhai',
@@ -25,15 +22,13 @@ describe('POST API Todo /post/todo', () => {
                 author: 'mrguri',
                 app_id: '6025e9b0bcfcf3098888eadc'
             })
-            .end((err, res) => {
-                expect(res.status).to.equal(200)
-            })
+        expect(res.status).toEqual(200)
     });
 })
 
 describe('PUT API Todo /put/todo', () => {
     it('update todo info', async () => {
-        request
+        let res = await request
             .put('/api/put/todo')
             .send({
                 title: 'guribhai',
@@ -41,21 +36,17 @@ describe('PUT API Todo /put/todo', () => {
                 author: 'KLS01',
                 todo_id: '602689c86dfb7c7ded72c200'
             })
-            .end((err, res) => {
-                expect(res.status).to.equal(200)
-            })
+        expect(res.status).toEqual(200)
     });
 })
 
 describe('DELETE Todo  /delete/todo', () => {
     it('delete todo with todo id', async () => {
-        request
+        let res = await request
             .delete('/api/delete/todo')
             .send({
                 todo_id: '602689c86dfb7c7ded72c200'
             })
-            .end((err, res) => {
-                expect(res.status).to.equal(200)
-            })
+        expect(res.status).toEqual(200)
     });
 })
