+++ server/src/__tests__/utils.test.js
@@ -1,19 +1,13 @@
 //Require the dev-dependencies
-import chai from 'chai';
-const expect = chai.expect;
-import server from '../../app.js';
 import supertest from 'supertest';
+import server from '../../app.js';
+import expect from 'expect';
 const request = supertest(server)
 
 describe('GET health check API /fail-health', () => {
     it('Failed Health Check', async () => {
-        request
+        let res = await request
             .get('/fail-health')
-            .end((err, res) => {
-                expect(res.status).to.equal(500)
-            })
+        expect(res.status).toEqual(500)
     });
 })
-
-
-
